requirements.txt
PyYAML==6.0.2
requests==2.32.3
cryptography==45.0.6
alibabacloud-cas20200407==1.2.8
alibabacloud-tea-openapi==0.4.1
tencentcloud-sdk-python-ssl==3.0.1329
